server/fastapidrone/requirements.txt
matplotlib==3.3.2
numpy==1.19.5
keras_segmentation==0.3.0
uvicorn==0.14.0
starlette==0.14.2
fastapi==0.68.1
Pillow==8.3.2
uvicorn==0.14.0
pydantic==1.8.2
tensorflow-cpu
tensorflow
Keras
gunicorn==19.9.0
h5py
opencv-python-headless
skia-python
python-multipart